// frontend-angular/src/app/core/generated-api/model/consultorioResponse.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Consultorio fisico donde se atienden las citas medicas
 */
export interface ConsultorioResponse { 
    readonly id: number;
    nombre: string;
    ubicacion: string;
    activo: boolean;
}