// frontend-angular/src/app/core/generated-api/model/diaSemana.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Dia de la semana para la disponibilidad horaria de un medico
 */
export const DiaSemana = {
    Lunes: 'Lunes',
    Martes: 'Martes',
    Miercoles: 'Miercoles',
    Jueves: 'Jueves',
    Viernes: 'Viernes',
    Sabado: 'Sabado',
} as const;
export type DiaSemana = typeof DiaSemana[keyof typeof DiaSemana];