// frontend-angular/src/app/core/generated-api/model/medico.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Medico { 
    readonly id: number;
    nombreCompleto: string;
    documentoIdentidad: string;
    especialidad: string;
    numeroColegiado: string;
    telefono?: string;
    email?: string;
    activo?: boolean;
}